+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Update DynamoCore to 2.8.0 for Dynamo CPython3 engine.
 * Add some View Nodes - View.HideCategoriesTemporary, View.HideElementsTemporary, View.IsolateCategoriesTemporary, View.IsolateElementsTemporary.
 * Improve DuplicateSheet Node - add "duplicateWithContents" option, add default suffix value when prefix & suffix are both empty.
+* Improve DuplicateSheet Node - Set Sheet information when duplicating; Improve View Temporary Nodes.
 
 ## 0.2.18
 * Add a Category ScheduleOnSheet and its nodes - ScheduleOnSheet.Sheet, ScheduleOnSheet.Schedule, ScheduleOnSheet.BySheetViewLocation, ScheduleOnSheet.Location and ScheduleOnSheet.SetLocation